chore: move cors config to utils
